+++ plone/app/event/ical/exporter.py
@@ -1,11 +1,11 @@
 from Acquisition import aq_inner
-from Products.ZCatalog.interfaces import ICatalogBrain
 from datetime import datetime
 from datetime import timedelta
 from plone.app.contentlisting.interfaces import IContentListingObject
-from plone.app.event.base import RET_MODE_BRAINS
 from plone.app.event.base import default_timezone
 from plone.app.event.base import get_events
+from plone.app.event.base import RET_MODE_BRAINS
+from plone.event.interfaces import IEvent
 from plone.event.interfaces import IEventAccessor
 from plone.event.interfaces import IICalendar
 from plone.event.interfaces import IICalendarEventComponent
@@ -13,6 +13,7 @@ from plone.event.interfaces import IOccurrence
 from plone.event.utils import is_datetime
 from plone.event.utils import tzdel
 from plone.event.utils import utc
+from Products.ZCatalog.interfaces import ICatalogBrain
 from zope.interface import implementer
 from zope.publisher.browser import BrowserView
 
@@ -49,6 +50,9 @@ def construct_icalendar(context, events):
         if ICatalogBrain.providedBy(event) or\
                 IContentListingObject.providedBy(event):
             event = event.getObject()
+        if not (IEvent.providedBy(event) or IOccurrence.providedBy(event)):
+            # Must be an event.
+            continue
         acc = IEventAccessor(event)
         tz = acc.timezone
         # TODO: the standard wants each recurrence to have a valid timezone
@@ -195,35 +199,61 @@ class ICalendarEventComponent(object):
 
     def __init__(self, context):
         self.context = context
-        self.event = IEventAccessor(context)
+        self.event = IEventAccessor(self.context)
+        self.ical = icalendar.Event()
 
-    def to_ical(self):
-        ical = icalendar.Event()
-        event = self.event
+    @property
+    def dtstamp(self):
+        # must be in uc
+        return {'value': utc(datetime.now())}
 
-        # TODO: event.text
+    @property
+    def created(self):
+        # must be in uc
+        return {'value': utc(self.event.created)}
 
-        # must be in utc
-        ical.add('dtstamp', utc(datetime.now()))
-        ical.add('created', utc(event.created))
-        ical.add('last-modified', utc(event.last_modified))
+    @property
+    def last_modified(self):
+        # must be in uc
+        return {'value': utc(self.event.last_modified)}
 
-        if event.sync_uid:
-            # Re-Use existing icalendar event UID
-            ical.add('uid', event.sync_uid)
-        else:
-            # Else, use plone.uuid
-            ical.add('uid', event.uid)
+    @property
+    def uid(self):
+        return {'value': self.event.sync_uid}
 
-        ical.add('url', event.url)
+    @property
+    def url(self):
+        return {'value': self.event.url}
 
-        ical.add('summary', event.title)
+    @property
+    def summary(self):
+        return {'value': self.event.title}
 
-        if event.description:
-            ical.add('description', event.description)
+    @property
+    def description(self):
+        return {'value': self.event.description}
 
-        if event.whole_day:
-            ical.add('dtstart', event.start.date())
+    @property
+    def dtstart(self):
+        if self.event.whole_day:
+            # RFC5545, 3.6.1
+            # For cases where a "VEVENT" calendar component
+            # specifies a "DTSTART" property with a DATE value type but no
+            # "DTEND" nor "DURATION" property, the event's duration is taken to
+            # be one day.
+            return {'value': self.event.start.date()}
+
+        # Normal case + Open End case:
+        # RFC5545, 3.6.1
+        # For cases where a "VEVENT" calendar component
+        # specifies a "DTSTART" property with a DATE-TIME value type but no
+        # "DTEND" property, the event ends on the same calendar date and
+        # time of day specified by the "DTSTART" property.
+        return {'value': self.event.start}
+
+    @property
+    def dtend(self):
+        if self.event.whole_day:
             # RFC5545, 3.6.1
             # For cases where a "VEVENT" calendar component
             # specifies a "DTSTART" property with a DATE value type but no
@@ -244,55 +274,76 @@ class ICalendarEventComponent(object):
             # -appointments-in-ics-files
             # http://icalevents.com/1778-all-day-events-adding-a-day-or-not/
             # http://www.innerjoin.org/iCalendar/all-day-events.html
-            ical.add('dtend', event.end.date() + timedelta(days=1))
-        elif event.open_end:
+            return {'value': self.event.end.date() + timedelta(days=1)}
+
+        elif self.event.open_end:
             # RFC5545, 3.6.1
             # For cases where a "VEVENT" calendar component
             # specifies a "DTSTART" property with a DATE-TIME value type but no
             # "DTEND" property, the event ends on the same calendar date and
             # time of day specified by the "DTSTART" property.
-            ical.add('dtstart', event.start)
-        else:
-            ical.add('dtstart', event.start)
-            ical.add('dtend', event.end)
+            return None
 
-        if event.recurrence and not IOccurrence.providedBy(self.context):
-            for recdef in event.recurrence.split():
-                prop, val = recdef.split(':')
-                if prop == 'RRULE':
-                    ical.add(prop, icalendar.prop.vRecur.from_ical(val))
-                elif prop in ('EXDATE', 'RDATE'):
-                    factory = icalendar.prop.vDDDLists
+        return {'value': self.event.end}
 
-                    # localize ex/rdate
-                    # TODO: should better already be localized by event object
-                    tzid = event.timezone
-                    if isinstance(tzid, tuple):
-                        tzid = tzid[0]
-                    # get list of datetime values from ical string
-                    try:
-                        dtlist = factory.from_ical(val, timezone=tzid)
-                    except ValueError:
-                        # TODO: caused by a bug in plone.formwidget.recurrence,
-                        # where the recurrencewidget or plone.event fails with
-                        # COUNT=1 and a extra RDATE.
-                        # TODO: REMOVE this workaround, once this failure is
-                        # fixed in recurrence widget.
-                        continue
-                    ical.add(prop, dtlist)
+    @property
+    def recurrence(self):
+        if not self.event.recurrence or IOccurrence.providedBy(self.context):
+            return None
 
-        if event.location:
-            ical.add('location', event.location)
+        ret = []
+        for recdef in self.event.recurrence.split():
+            prop, val = recdef.split(':')
+            if prop == 'RRULE':
+                ret.append({
+                    'property': prop,
+                    'value': icalendar.prop.vRecur.from_ical(val)
+                })
 
+            elif prop in ('EXDATE', 'RDATE'):
+                factory = icalendar.prop.vDDDLists
+
+                # localize ex/rdate
+                # TODO: should better already be localized by event object
+                tzid = self.event.timezone
+                if isinstance(tzid, tuple):
+                    tzid = tzid[0]
+                # get list of datetime values from ical string
+                try:
+                    dtlist = factory.from_ical(val, timezone=tzid)
+                except ValueError:
+                    # TODO: caused by a bug in plone.formwidget.recurrence,
+                    # where the recurrencewidget or plone.event fails with
+                    # COUNT=1 and a extra RDATE.
+                    # TODO: REMOVE this workaround, once this failure is
+                    # fixed in recurrence widget.
+                    continue
+                ret.append({
+                    'property': prop,
+                    'value': dtlist
+                })
+
+        return ret
+
+    @property
+    def location(self):
+        return {'value': self.event.location}
+
+    @property
+    def attendee(self):
         # TODO: revisit and implement attendee export according to RFC
-        if event.attendees:
-            for attendee in event.attendees:
-                att = icalendar.prop.vCalAddress(attendee)
-                att.params['cn'] = icalendar.prop.vText(attendee)
-                att.params['ROLE'] = icalendar.prop.vText('REQ-PARTICIPANT')
-                ical.add('attendee', att)
+        ret = []
+        for attendee in self.event.attendees or []:
+            att = icalendar.prop.vCalAddress(attendee)
+            att.params['cn'] = icalendar.prop.vText(attendee)
+            att.params['ROLE'] = icalendar.prop.vText('REQ-PARTICIPANT')
+            ret.append(att)
+        return {'value': ret}
 
+    @property
+    def contact(self):
         cn = []
+        event = self.event
         if event.contact_name:
             cn.append(event.contact_name)
         if event.contact_phone:
@@ -301,14 +352,59 @@ class ICalendarEventComponent(object):
             cn.append(event.contact_email)
         if event.event_url:
             cn.append(event.event_url)
-        if cn:
-            ical.add('contact', u', '.join(cn))
 
-        if event.subjects:
-            for subject in event.subjects:
-                ical.add('categories', subject)
+        return {'value': u', '.join(cn)}
 
-        return ical
+    @property
+    def categories(self):
+        ret = []
+        for cat in self.event.subjects or []:
+            ret.append({'value': cat})
+        return ret or None
+
+    @property
+    def geo(self):
+        """Not implemented.
+        """
+        return
+
+    def ical_add(self, prop, val):
+        if not val:
+            return
+
+        if not isinstance(val, list):
+            val = [val]
+
+        for _val in val:
+            assert(isinstance(_val, dict))
+            value = _val['value']
+            if not value:
+                continue
+            prop = _val.get('property', prop)
+            params = _val.get('parameters', None)
+            self.ical.add(prop, value, params)
+
+    def to_ical(self):
+        # TODO: event.text
+
+        ical_add = self.ical_add
+        ical_add('dtstamp', self.dtstamp)
+        ical_add('created', self.created)
+        ical_add('last-modified', self.last_modified)
+        ical_add('uid', self.uid)
+        ical_add('url', self.url)
+        ical_add('summary', self.summary)
+        ical_add('description', self.description)
+        ical_add('dtstart', self.dtstart)
+        ical_add('dtend', self.dtend)
+        ical_add(None, self.recurrence)  # property key set via val
+        ical_add('location', self.location)
+        ical_add('attendee', self.attendee)
+        ical_add('contact', self.contact)
+        ical_add('categories', self.categories)
+        ical_add('geo', self.geo)
+
+        return self.ical
 
 
 class EventsICal(BrowserView):
@@ -320,10 +416,13 @@ class EventsICal(BrowserView):
         return cal.to_ical()
 
     def __call__(self):
-        name = '%s.ics' % self.context.getId()
-        self.request.RESPONSE.setHeader('Content-Type', 'text/calendar')
-        self.request.RESPONSE.setHeader(
+        ical = self.get_ical_string()
+        name = '{0}.ics'.format(self.context.getId())
+        self.request.response.setHeader('Content-Type', 'text/calendar')
+        self.request.response.setHeader(
             'Content-Disposition',
-            'attachment; filename="%s"' % name
+            'attachment; filename="{0}"'.format(name)
         )
-        self.request.RESPONSE.write(self.get_ical_string())
+        self.request.response.setHeader('Content-Length', len(ical))
+        self.request.response.setHeader('Pragma', 'no-cache')
+        self.request.response.write(ical)